Bugfix: task completion block never called when not following redirects.
When `-[NSURLSessionTaskDelegate URLSession:task:willPerformHTTPRedirection:newRequest:completionHandler:]` calls its completion handler with `nil` to indicate that the redirect should not be followed, the completion handler from `-[NSURLSession dataTaskWithRequest:completionHandler:]` is normally called with the actual redirect response itself. But when using OHTTPStubs such that a stub returns a redirect response, the task’s completion handler block is simply never called.

- Ensure `OHHTTPStubsProtocol` finishes redirect responses fully when not following redirects.
- Update `NSURLSessionTests` to exercise this behaviour.
- Fix `AFNetworkingTests` test that had been looking for the incorrect behaviour.