Fix redirect race condition
The redirect and response finished callbacks were being executed in two
separate blocks, each submitted using dispatch_after. The wait time for
each dispatch_after block was equal (and potentially 0 if the user did
not configure a request time). As a result, the order the blocks ran in
was non-deterministic and unpredictable. This broke the strong ordering
of the callbacks. When the response finished callback arrived before the
redirect callback, the OHHTTPStubs client would be given the contents of
the 3xx response with no opportunity to follow the redirect.

By moving both callbacks into a single block, strong ordering is again
guaranteed and redirect callbacks always happen when appropriate.

For more details about the required ordering of NSURLProtocolClient
callbacks, see the "Threading Notes" section of the readme of the Apple
CustomHTTPProtocol sample code project:

https://developer.apple.com/library/content/samplecode/CustomHTTPProtocol/Listings/Read_Me_About_CustomHTTPProtocol_txt.html